DHFPROD-10455: Consume basepath in HC UI and navigate
